+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Moved Validator and validation_class_iter to ingest-validation-tests
 - Deprecated cleanup examples
 - Catching entity-api exceptions
+- Bugfix multi-assay validation, error if no required components found
 
 ## v0.0.39
 - Create Object-by-Analyte EPIC directory schema